+++ AP3/TP/include/fibonacci.h
@@ -3,7 +3,7 @@
 
 #include <iostream>
 
-int fibonacciIterativo(int n);
-int fibonacciRecursivo(int n);
+long double fibonacciIterativo(long double n);
+long double fibonacciRecursivo(long double n);
 
 #endif //FIBONACCI_H
